+++ gecko_client/health_api.go
@@ -2,6 +2,7 @@ package gecko_client
 
 import (
 	"encoding/json"
+	"github.com/kurtosis-tech/ava-e2e-tests/gecko_client/rpc_requester"
 	"github.com/palantir/stacktrace"
 )
 
@@ -11,11 +12,11 @@ const (
 
 
 type HealthApi struct {
-	rpcRequester jsonRpcRequester
+	rpcRequester rpc_requester.JsonRpcRequester
 }
 
 func (api HealthApi) GetLiveness() (LivenessInfo, error) {
-	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(healthApiEndpoint, "health.getLiveness", make(map[string]interface{}))
+	responseBodyBytes, err := api.rpcRequester.MakeRpcRequest(healthApiEndpoint, "health.getLiveness", make(map[string]interface{}))
 	if err != nil {
 		return LivenessInfo{}, stacktrace.Propagate(err, "Error getting liveness")
 	}