+++ gecko_client/keystore_api.go
@@ -2,6 +2,7 @@ package gecko_client
 
 import (
 	"encoding/json"
+	"github.com/kurtosis-tech/ava-e2e-tests/gecko_client/rpc_requester"
 	"github.com/palantir/stacktrace"
 )
 
@@ -10,7 +11,7 @@ const (
 )
 
 type KeystoreApi struct {
-	rpcRequester jsonRpcRequester
+	rpcRequester rpc_requester.JsonRpcRequester
 }
 
 // Creates a blockchain with the given parameters, returning the unsigned transaction identifier
@@ -19,7 +20,7 @@ func (api KeystoreApi) CreateUser(username string, password string) (bool, error
 		"username": username,
 		"password": password,
 	}
-	responseBodyBytes, err := api.rpcRequester.makeRpcRequest(keystoreEndpoint, "keystore.createUser", params)
+	responseBodyBytes, err := api.rpcRequester.MakeRpcRequest(keystoreEndpoint, "keystore.createUser", params)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "Error making request")
 	}