+++ gecko_client/rpc_requester/json_rpc_requester.go
@@ -1,0 +1,5 @@
+package rpc_requester
+
+type JsonRpcRequester interface {
+	MakeRpcRequest(endpoint string, method string, params map[string]interface{}) ([]byte, error)
+}